+++ jacoco-filtering-extension/src/main/kotlin/io/github/surpsg/deltacoverage/report/ReportsFactory.kt
@@ -1,8 +1,9 @@
 package io.github.surpsg.deltacoverage.report
 
+import io.github.surpsg.deltacoverage.config.CoverageEntity
+import io.github.surpsg.deltacoverage.config.CoverageRulesConfig
 import io.github.surpsg.deltacoverage.config.DeltaCoverageConfig
 import io.github.surpsg.deltacoverage.config.ReportsConfig
-import io.github.surpsg.deltacoverage.config.ViolationRuleConfig
 import io.github.surpsg.deltacoverage.diff.DiffSource
 import org.jacoco.core.analysis.ICoverageNode
 import org.jacoco.report.check.Limit
@@ -15,17 +16,14 @@ internal fun reportFactory(
 ): Set<FullReport> {
     val reports: Set<Report> = diffSourceConfig.reportsConfig.toReportTypes()
 
-    val violationRule: Rule = buildRule(diffSourceConfig.violationRuleConfig)
+    val rules: List<Rule> = diffSourceConfig.coverageRulesConfig.buildRules()
     val baseReportDir = Paths.get(diffSourceConfig.reportsConfig.baseReportDir)
     val report: MutableSet<FullReport> = mutableSetOf(
         DiffReport(
             baseReportDir.resolve("deltaCoverage"),
             reports,
             diffSource,
-            Violation(
-                diffSourceConfig.violationRuleConfig.failOnViolation,
-                listOf(violationRule)
-            )
+            Violation(diffSourceConfig.coverageRulesConfig.failOnViolation, rules)
         )
     )
 
@@ -47,23 +45,33 @@ private fun ReportsConfig.toReportTypes(): Set<Report> = sequenceOf(
     Report(it.first, it.second.outputFileName)
 }.toSet()
 
-private fun buildRule(
-    violationRulesOptions: ViolationRuleConfig
-): Rule {
-    return sequenceOf(
-        ICoverageNode.CounterEntity.INSTRUCTION to violationRulesOptions.minInstructions,
-        ICoverageNode.CounterEntity.BRANCH to violationRulesOptions.minBranches,
-        ICoverageNode.CounterEntity.LINE to violationRulesOptions.minLines
-    ).filter {
-        it.second > 0.0
-    }.map {
-        Limit().apply {
-            setCounter(it.first.name)
-            minimum = it.second.toString()
+private fun CoverageRulesConfig.buildRules(): List<Rule> {
+    val limits: List<Limit> = entitiesRules.asSequence()
+        .map { (coverageEntity, minValue) ->
+            coverageEntity.toJacocoEntity() to minValue.minCoverageRatio
         }
-    }.toList().let {
-        Rule().apply {
-            limits = it
-        }
+        .filter { (_, minCoverage) ->
+            minCoverage > 0.0
+        }.map { (counterType, minCoverage) ->
+            Limit().apply {
+                setCounter(counterType.name)
+                minimum = minCoverage.toString()
+            }
+        }.toList()
+
+    return if (limits.isNotEmpty()) {
+        listOf(
+            Rule().apply { this.limits = limits }
+        )
+    } else {
+        emptyList()
+    }
+}
+
+private fun CoverageEntity.toJacocoEntity(): ICoverageNode.CounterEntity {
+    return when (this) {
+        CoverageEntity.INSTRUCTION -> ICoverageNode.CounterEntity.INSTRUCTION
+        CoverageEntity.BRANCH -> ICoverageNode.CounterEntity.BRANCH
+        CoverageEntity.LINE -> ICoverageNode.CounterEntity.LINE
     }
 }