+++ jacoco-filtering-extension/src/test/kotlin/io/github/surpsg/deltacoverage/coverage/diff/DiffSourceFactoryTest.kt
@@ -11,6 +11,8 @@ import io.kotest.matchers.should
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.startWith
 import io.kotest.matchers.types.shouldBeTypeOf
+import io.mockk.every
+import io.mockk.mockk
 import java.io.File
 
 class DiffSourceFactoryTest : StringSpec() {
@@ -19,7 +21,7 @@ class DiffSourceFactoryTest : StringSpec() {
         "diffSourceFactory should return file diff source" {
             // setup
             val filePath = "someFile"
-            val diffConfig = buildDiffSourceOption(file = filePath)
+            val diffConfig = DiffSourceConfig { file = filePath }
 
             // run
             val diffSource = diffSourceFactory(File("."), diffConfig)
@@ -31,21 +33,21 @@ class DiffSourceFactoryTest : StringSpec() {
 
         "diffSourceFactory should return url diff source" {
             // setup
-            val url = "someUrl"
-            val diffConfig = buildDiffSourceOption(url = url)
+            val expectedUrl = "someUrl"
+            val diffConfig = DiffSourceConfig { url = expectedUrl }
 
             // run
             val diffSource = diffSourceFactory(File("."), diffConfig)
 
             // assert
             diffSource.shouldBeTypeOf<UrlDiffSource>()
-            diffSource.sourceDescription shouldBe "URL: $url"
+            diffSource.sourceDescription shouldBe "URL: $expectedUrl"
         }
 
         "diffSourceFactory should return git diff source" {
             // setup
             val compareWith = "develop"
-            val diffConfig = buildDiffSourceOption(git = compareWith)
+            val diffConfig = DiffSourceConfig { diffBase = compareWith }
 
             // run
             val diffSource = diffSourceFactory(File("."), diffConfig)
@@ -57,8 +59,11 @@ class DiffSourceFactoryTest : StringSpec() {
 
         "diffSourceFactory should throw when no source specified" {
             // setup
-            val diffConfig = buildDiffSourceOption()
-
+            val diffConfig = mockk<DiffSourceConfig> {
+                every { file } returns ""
+                every { url } returns ""
+                every { diffBase } returns ""
+            }
             // run
             val exception = shouldThrow<IllegalStateException> {
                 diffSourceFactory(File("."), diffConfig)
@@ -69,24 +74,5 @@ class DiffSourceFactoryTest : StringSpec() {
                 "Expected Git configuration or file or URL diff source but all are blank"
             )
         }
-
-        "diffSourceFactory should throw when both sources specified" {
-            // setup
-            val diffConfig = buildDiffSourceOption(file = "file", url = "url", git = "master")
-
-            // run
-            val exception = shouldThrow<IllegalStateException> {
-                diffSourceFactory(File("."), diffConfig)
-            }
-
-            // assert
-            exception.message should startWith(
-                "Expected only Git configuration or file or URL diff source more than one:"
-            )
-        }
-    }
-
-    private fun buildDiffSourceOption(file: String = "", url: String = "", git: String = ""): DiffSourceConfig {
-        return DiffSourceConfig(file, url, git)
     }
 }